fix: use modal instead of new page
