+++ client/src/components/Subscription/index.ts
@@ -3,4 +3,6 @@ export { default as PaymentForm } from './PaymentForm';
 export { default as PlanSummary } from './PlanSummary';
 export { default as PriceBreakdown } from './PriceBreakdown';
 export { default as SubscribePrompt } from './SubscribePrompt';
+export { default as ChoosePlanModal } from './ChoosePlanModal';
+export { default as CheckoutModal } from './CheckoutModal';
 export type { PlanCardProps } from './PlanCard';
