+++ client/src/routes/Root.tsx
@@ -20,6 +20,7 @@ import { TermsAndConditionsModal } from '~/components/ui';
 import { Nav, MobileNav } from '~/components/Nav';
 import { useHealthCheck } from '~/data-provider';
 import { Banner } from '~/components/Banners';
+import { ChoosePlanModal, CheckoutModal } from '~/components/Subscription';
 
 export default function Root() {
   const [showTerms, setShowTerms] = useState(false);
@@ -92,6 +93,8 @@ export default function Root() {
               modalContent={config.interface.termsOfService.modalContent}
             />
           )}
+          <ChoosePlanModal />
+          <CheckoutModal />
         </AssistantsMapContext.Provider>
       </FileMapContext.Provider>
     </SetConvoProvider>
