+++ client/src/store/settings.ts
@@ -15,6 +15,9 @@ const staticAtoms = {
     default: SettingsViews.default,
   }),
   showPopover: atom<boolean>({ key: 'showPopover', default: false }),
+  showChoosePlan: atom<boolean>({ key: 'showChoosePlan', default: false }),
+  showCheckout: atom<boolean>({ key: 'showCheckout', default: false }),
+  checkoutPlanId: atom<string | null>({ key: 'checkoutPlanId', default: null }),
 };
 
 const localStorageAtoms = {
